+++ EngineHacks/CoreHacks/StatusGetter/Ranks/ranks.c
@@ -2,13 +2,13 @@
 #include "StatusGetter.h"
 
 // W.I.P
-int GetRankReason(Unit* unit){
+s8 GetRankReason(Unit* unit){
 	
 	return TH_CLASS_S;
 }
 
 
-int GetRankFaith(Unit* unit){
+s8 GetRankFaith(Unit* unit){
 	
 	return TH_CLASS_S;
 }
